SearchQueryModel created

+++ BackEnd/MyWebApi/Models/Edge.cs
@@ -4,7 +4,7 @@ namespace MyWebApi.Models
     {
         public string SourceAcount;
         public string DestiantionAccount; //recursively sending all the component alert so use ToString methode
-        public int Amount;
+        public long Amount;
         public string Date;
         public long TransactionId;
         public string Type;
